refactor: simplify test cases and remove unused variables in integration tests
